Add a -quiet flag
Some projects will have a very large number of dependencies installed,
and the default behaviour prints a message for each package whether or
not it has vulnerabilities. This adds a -quiet flag that would print
output for packages only if they have vulnerabilities, to make it
easier to see at a glance what issues exist. The summary information,
packages analyzed vs vulnerable, is always printed.