Add integration test to validate the results are the same
Adds a sample Gopkg.lock and Gopkg.toml. This will be needed when we
switch over to dep. And what better project to use as a test than the
thign we will be using :)

All the setup work atm is more dep specific since it looks like it needs
to be fully setup like a real GOPATH.