Making it so that CVE exclude can be read from a file
Default to a file named exclude_vulnerabilities that will live at the
root of whereever you are running this. The flag of
'exclude-vulnerability-file' will override whatever path you want it to
be.

Also you can use it in conjunction with exclude-vulnerability and they
will be merged together if you are into that kind of thing.