Update version of aurora to fix go get
Running this as i understand it
go get -u github.com/sonatype-nexus-community/nancy

Any depenency that is define is not used unless the dependency is
included directly in your project.
https://github.com/golang/dep/blob/master/docs/FAQ.md#why-is-dep-ignoring-a-version-constraint-in-the-manifest

So basically it was pulling latest master of aurora and they had changed
the method signature of Gray.

To fix this just simply updated aurora and updated the one method call
to Gray