Allow go list -m all via std in  (#43)
As it says, allows for sending in go list -m all via std in. Also since i had to tweaked the code and initially broke when usage would be printed if nothing was passed, i figured a nancy -help was in order but then realized I missed the location I actually needed to put the new usage call. So now it functions like it originally did but i left the new help flag b/c it cant hurt.