Adding a way to setup expiring ignores.
Once the until date is passed todays date it will no longer be added to
the exclusion list and will force it to be dealt with at that time.

Pattern is as follows

CVE-11 until=2019-01-01 #Ignoring this for now ill revisit after january

It's relatively flexible as far as spacing goes and even comments go.
But generally what is above is the desired and probably most readable
format.