Clean up dep artifacts from project root
Having this and go.mod at the root is confusing. We only left them to
use for test. There are some in the testdata/dep directory so using
those as part of CI to make sure it still works.

That way new contributors can clearly look at the nancy project and
know its using go mod.