This should fix your local @bhamail
Issue with running circle ci locally is it requires docker in docker.
That no worky. See here
https://github.com/CircleCI-Public/circleci-cli/issues/291

You can cheat and run it as root but gross.

So instead of that used this advice found here
https://stackoverflow.com/a/45796662

Where we now check what env you are doing and if CIRCLE_SHELL_ENV
contains localbuild. We will sudo with docker to make it work. Otherwise
we just do the normal docker.